refactor: 폴더 위치 변경 및 Input 테마 적용
