+++ frontend/techpick/src/features/nodeManagement/api/useDeleteFolder.ts
